Move `BloodSugarValidator#ValidationResult` to its own file
