+++ app/src/main/java/org/simple/clinic/bloodsugar/entry/BloodSugarValidator.kt
@@ -2,21 +2,13 @@ package org.simple.clinic.bloodsugar.entry
 
 import org.simple.clinic.bloodsugar.BloodSugarMeasurementType
 import org.simple.clinic.bloodsugar.HbA1c
-import org.simple.clinic.bloodsugar.entry.BloodSugarValidator.ValidationResult.ErrorBloodSugarEmpty
-import org.simple.clinic.bloodsugar.entry.BloodSugarValidator.ValidationResult.ErrorBloodSugarTooHigh
-import org.simple.clinic.bloodsugar.entry.BloodSugarValidator.ValidationResult.ErrorBloodSugarTooLow
-import org.simple.clinic.bloodsugar.entry.BloodSugarValidator.ValidationResult.Valid
+import org.simple.clinic.bloodsugar.entry.ValidationResult.ErrorBloodSugarEmpty
+import org.simple.clinic.bloodsugar.entry.ValidationResult.ErrorBloodSugarTooHigh
+import org.simple.clinic.bloodsugar.entry.ValidationResult.ErrorBloodSugarTooLow
+import org.simple.clinic.bloodsugar.entry.ValidationResult.Valid
 import javax.inject.Inject
 
 class BloodSugarValidator @Inject constructor() {
-
-  sealed class ValidationResult {
-    data class Valid(val bloodSugarReading: Float) : ValidationResult()
-
-    object ErrorBloodSugarEmpty : ValidationResult()
-    data class ErrorBloodSugarTooHigh(val measurementType: BloodSugarMeasurementType) : ValidationResult()
-    data class ErrorBloodSugarTooLow(val measurementType: BloodSugarMeasurementType) : ValidationResult()
-  }
 
   fun validate(bloodSugarReading: String, measurementType: BloodSugarMeasurementType): ValidationResult {
     if (bloodSugarReading.isBlank()) {
